dom/tests/mochitest/ajax/offline/test_overlap.html

<html xmlns="http://www.w3.org/1999/xhtml" manifest="http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/overlap.cacheManifest">
<head>
<title>overlapping namespaces test</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="/tests/dom/tests/mochitest/ajax/offline/offlineTests.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

function manifestUpdated()
{
  var xhr = new XMLHttpRequest();
  xhr.open("GET", "bogus/specific/test.html", false);
  xhr.send();

  OfflineTest.is(xhr.status, 200, "Should have fallen back.");
  OfflineTest.teardown();
  OfflineTest.finish();
}

if (OfflineTest.setup()) {
  applicationCache.onerror = OfflineTest.failEvent;
  applicationCache.oncached = OfflineTest.priv(manifestUpdated);
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body>

</body>
</html>
